+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+## [1.10.3-beta.1]
 ### enhancements
 - add debugger cli toggles to tracing more info on logs.
 - media modules now exports the app related to the media player.